+++ changelog.md
@@ -1,4 +1,8 @@
 # v3.0.0
 
+- Added option to parse graph from a list of instructions
 - Removed debug feature
+
+## Bugfixes
+
 - Fixed dijkstra implementation to also compute the correct path when edges are already in the open list but with a larger distance